Refactor: ApplyService 생성 후 addApplyIndex 메서드 이동
