+++ src/main/java/com/mycompany/metamong/service/ApplyService.java
@@ -1,0 +1,25 @@
+package com.mycompany.metamong.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.mycompany.metamong.daoMain.IndexDao;
+import com.mycompany.metamong.dto.applyList.ApplyListDto;
+import com.mycompany.metamong.dto.index.ApplyIndexDto;
+
+@Service
+public class ApplyService {
+	@Autowired
+	private ApplyListService applyListService;
+	@Autowired
+	private IndexDao indexDao;
+	
+	@Transactional
+	public void addApplyIndex(ApplyListDto applyListDto, ApplyIndexDto applyIndexDto) {
+		applyListService.addApplyList(applyListDto);
+		applyIndexDto.setApplyNo(applyListDto.getApplyNo());
+		indexDao.insertApplyIndex(applyIndexDto);
+	}
+	
+}
